% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autodatadoc-package.R
\docType{package}
\name{autodatadoc-package}
\alias{autodatadoc}
\alias{autodatadoc-package}
\title{autodatadoc: Example of automated data documentation for R packages}
\description{
Example of automated data documentation for R packages showing how to store a data dictionary externally from R source code yet have it automatically synced with roxygen documentation.
}
\author{
\strong{Maintainer}: Joseph Stachelek \email{stachel2@msu.edu} (\href{https://orcid.org/0000-0002-5924-2464}{ORCID})

}
\keyword{internal}
