% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{World Health Organization TB data}
\format{
\tabular{ll}{
\bold{name} \tab \bold{description} \cr
country    \tab Country name\cr
  year       \tab Year        \cr
  population \tab Population  
}
}
\usage{
population
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report, and accompanying global populations.
}
\keyword{datasets}
